requirements.txt
python-dotenv
azure-cosmos
streamlit